api/requirements.txt
Django>=5.2.4,<5.3
djangorestframework>=3.16.0,<3.17

djangorestframework-simplejwt>=5.5.0,<5.6
django-oauth-toolkit>=3.0.1,<3.1

psycopg2-binary>=2.9.3,<2.10

pillow>=11.3.0,<11.4
drf-spectacular>=0.28.0,<0.29

pyjwt>=2.10.1,<2.11
python-dotenv>=1.1.1,<1.2

celery>=5.5.3,<6.0
redis>=6.0.0,<7.0